requirements.txt
# Core processing dependencies
pyspark==3.5.0
apache-beam[gcp]==2.52.0

# Database connections
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Kafka integration
kafka-python==2.0.2
confluent-kafka==2.3.0

# HTTP and API
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0

# Data validation and utilities
pydantic==2.5.0
python-dotenv==1.0.0
click==8.1.7

# Development tools
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring and logging
structlog==23.2.0